+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - [#639](https://github.com/helmholtz-analytics/heat/pull/639) Bugfix: balanced array in demo_knn, changed behaviour of knn
 - [#648](https://github.com/helmholtz-analytics/heat/pull/648) Bugfix: tensor printing with PyTorch 1.6.0
 - [#653](https://github.com/helmholtz-analytics/heat/pull/653) Update unittests argmax & argmin + force index order in mpi_argmax & mpi_argmin.
+- [#653](https://github.com/helmholtz-analytics/heat/pull/653) Printing above threshold gathers the data without a buffer now
 
 # v0.4.0